Explicit protocol in gravatar links
Although // links are better normally, it appears that within email bodies, they
are not processed like they would be in a web browser context. As a result, any
user with a Gravatar will not have their picture show up in emails. This commit
explicitly uses HTTPS so that Gravatars show up in emails.